% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_sbproj_view.R
\name{as_sbproj_view}
\alias{as_sbproj_view}
\title{Convert an object into an "sbproj_view" object}
\usage{
as_sbproj_view(x, ...)
}
\arguments{
\item{x}{The object to convert}

\item{...}{Ignored}
}
\value{
An "sbproj_view" object
}
\description{
Convert an object into an "sbproj_view" object
}
