% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_sbproj_view.R
\name{assert_xml_attr_names}
\alias{assert_xml_attr_names}
\title{Verify that an XML node has given attribute names}
\usage{
assert_xml_attr_names(x, ...)
}
\arguments{
\item{x}{The xml2 node object}

\item{...}{
  Arguments passed on to \code{\link[checkmate:checkNames]{checkmate::check_names}}
  \describe{
    \item{\code{type}}{[character(1)]\cr
Type of formal check(s) to perform on the names.
\describe{
\item{unnamed:}{Checks \code{x} to be \code{NULL}.}
\item{named:}{Checks \code{x} for regular names which excludes names to be \code{NA} or empty (\code{""}).}
\item{unique:}{Performs checks like with \dQuote{named} and additionally tests for non-duplicated names.}
\item{strict:}{Performs checks like with \dQuote{unique} and additionally fails for names with UTF-8 characters and names which do not comply to R's variable name restrictions.
  As regular expression, this is \dQuote{^[.]*[a-zA-Z]+[a-zA-Z0-9._]*$}.}
\item{ids:}{Same as \dQuote{strict}, but does not enforce uniqueness.}
}
Note that for zero-length \code{x}, all these name checks evaluate to \code{TRUE}.}
    \item{\code{subset.of}}{[\code{character}]\cr
Names provided in \code{x} must be subset of the set \code{subset.of}.}
    \item{\code{must.include}}{[\code{character}]\cr
Names provided in \code{x} must be a superset of the set \code{must.include}.}
    \item{\code{permutation.of}}{[\code{character}]\cr
Names provided in \code{x} must be a permutation of the set \code{permutation.of}.
Duplicated names in \code{permutation.of} are stripped out and duplicated names in \code{x}
thus lead to a failed check.
Use this argument instead of \code{identical.to} if the order of the names is not relevant.}
    \item{\code{identical.to}}{[\code{character}]\cr
Names provided in \code{x} must be identical to the vector \code{identical.to}.
Use this argument instead of \code{permutation.of} if the order of the names is relevant.}
    \item{\code{disjunct.from}}{[\code{character}]\cr
Names provided in \code{x} must may not be present in the vector \code{disjunct.from}.}
    \item{\code{what}}{[\code{character(1)}]\cr
Type of name vector to check, e.g. \dQuote{names} (default), \dQuote{colnames} or \dQuote{rownames}.}
  }}
}
\value{
x
}
\description{
Verify that an XML node has given attribute names
}
