% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_sbproj_view.R
\name{get_class_from_list}
\alias{get_class_from_list}
\title{Get an element that is a class from a list of objects}
\usage{
get_class_from_list(x, what, what_not = NULL, len = NULL)
}
\arguments{
\item{x}{The list to search}

\item{what}{The class name (passed to \code{inherts()})}

\item{what_not}{The class name(s) it should \emph{not} be}

\item{len}{The expected length of the return value (may be a vector of
values)}
}
\value{
A list of elements with the required class, or if
\code{all(len \%in\% c(0, 1))}, the value within the list element
}
\description{
Get an element that is a class from a list of objects
}
