% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sbproj.R
\name{read_sbproj}
\alias{read_sbproj}
\title{Read a Matlab .sbproj file}
\usage{
read_sbproj(file, ...)
}
\arguments{
\item{file}{The filename to read (usually ending in .sbproj)}

\item{...}{Passed to \code{read_sbproj_view()}}
}
\value{
An "sbproj" S3 object
}
\description{
Read a Matlab .sbproj file
}
