% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sbproj.R
\name{read_sbproj_view}
\alias{read_sbproj_view}
\title{Read the .view file within an .sbproj file}
\usage{
read_sbproj_view(file, ...)
}
\arguments{
\item{file}{The filename to read}

\item{...}{Passed to \code{as_sbproj_view()}}
}
\value{
An "sbproj_view" object
}
\description{
Read the .view file within an .sbproj file
}
